"no-op" libusnic_verbs: first implementation
Cisco has stopped supporting the Verbs API for general use.  However,
libibverbs will print a scary-looking stderr message if it finds a
usNIC device in /sys, but no corresponding userspace plugin for it.

Hence, we provide this userspace plugin for Cisco usNIC devices -- but
it provides no functionality.  It's sole purpose in life is to avoid
the libibverbs stderr warning.

(NOTE: Future versions of libibverbs will not emit this stderr
warning, but that doesn't help customers with older versions of
libibverbs -- so we created this empty plugin for legacy reasons).