CSCvh99543: update to 6-digit / version.sh scheme
Output files are now of the form:

* libusnic_verbs-a.b.c.d.tar.bz2
* libusnic_verbs-a.b.c.d.distro-e.f.src.rpm
* libusnic_verbs-a.b.c.d.distro-e.f.arch.rpm
* libusnic_verbs_a.b.c.d-e.f.distro_arch.changes
* libusnic_verbs_a.b.c.d-e.f.distro_arch.deb
* libusnic_verbs_a.b.c.d-e.f.distro.debian.tar.xz
* libusnic_verbs_a.b.c.d-e.f.distro.dsc

There is only a .bz2 tarball created now.  Also, there is no more
--build argument to the releng-* scripts -- the E.F build number is
assumed to be provided from version.sh, which will be [over]written
when releng does a build.

Releng wants to have *all* the digits (A-F) of the version number, so
there will likely be a follow-on commit to replace all of the
git-committed version.sh with A through F set to 0.  The current
values in version.sh reflect what releng's initial A and B values
should be.

Signed-off-by: Jeff Squyres <jsquyres@cisco.com>